Driving robot
You can now drive the robot forwards and backwards. You can rotate it
left and right. You can also set the speed for for the motors while
doing this or before doing it or maybe you want to stand still, then set
to rotate and then activate motors, you can do that.